requirements.txt
Babel==2.3.4
backports.functools-lru-cache==1.6.1
beautifulsoup4==4.8.1
braintree==3.59.0
boto3==1.17.112
botocore==1.20.112
certifi==2021.05.30
cffi==1.14.3
cloudpickle==1.2.2
cryptography==2.8
cycler==0.10.0
dateparser==0.7.2
decorator==4.0.10
docopt==0.6.2
docutils==0.16
ebaysdk==2.1.4
enum34==1.1.6
et-xmlfile==1.0.1
feedparser==5.2.1
funcsigs==1.0.2
futures==3.3.0
fuzzywuzzy==0.17.0
gevent==20.9.0
greenlet==0.4.17
ipaddress==1.0.23
jcconv==0.2.3
jdcal==1.4.1
Jinja2==2.8
jmespath==0.10.0
kiwisolver==1.1.0
lxml==4.2.5
Mako==1.0.4
MarkupSafe==0.23
matplotlib==2.2.4
mock==2.0.0
mws==0.8.9
networkx==2.2
num2words==0.5.10
numpy==1.16.5
OdooRPC==0.7.0
ofxparse==0.15
opencv-python==4.1.1.26
openpyxl==2.6.4
paramiko==2.7.2
passlib==1.6.5
paypalrestsdk==1.13.1
pbr==5.4.3
Pillow==3.4.1
psutil==5.7.2
psycogreen==1.0
psycopg2==2.8.6
psycopg2-binary==2.8.4
pycparser==2.19
pydot==1.2.3
PyJWT==1.7.1
PyMySQL==0.9.3
pyodbc==4.0.30
pyOpenSSL==19.0.0
pyotp==2.3.0
pyparsing==2.1.10
pyPdf==1.13
PyPDF2==1.26.0
PyPDF4==1.27.0
pyserial==3.1.1
Python-Chart==1.39
python-dateutil==2.5.3
python-magic==0.4.18
python-openid==2.2.5
python-stdnum==1.12
pytz==2016.7
pyusb==1.0.0
PyWavelets==1.0.3
PyYAML==3.12
qrcode==5.3
redis==3.5.3
regex==2019.11.1
reportlab==3.3.0
requests==2.11.1
revolut-python==0.6.1
scandir==1.10.0
scikit-image==0.14.5
scipy==1.2.2
selenium==3.141.0
simplejson==3.16.0
six==1.10.0
soupsieve==1.9.5
subprocess32==3.5.4
suds==0.4
suds-jurko==0.6
timeout-decorator==0.4.1
tzlocal==2.0.0
Unidecode==1.1.1
urllib3==1.25.6
validate-email==1.3
vatnumber==1.2
vobject==0.9.3
Wand==0.5.7
Werkzeug==0.11.11
xlrd==1.2.0
XlsxWriter==0.9.3
xlutils==2.0.0
xlwt==1.1.2
zope.event==4.5.0
zope.interface==5.1.2
pdfminer.six==20191110
